--- src/components/dDay/index.tsx
@@ -1,5 +1,0 @@
-const dDay = () => {
-  return <div></div>;
-};
-
-export default dDay;
