+++ src/components/layout/Header.tsx
@@ -1,8 +1,10 @@
+import JuiceFont from "../common/JuiceFont";
+
 const Header = () => {
   return (
     <div className="flex justify-between bg-stone-100 px-2 py-4">
-      <div className="font-mono font-bold">JUICE</div>
-      <div className="font-mono font-bold">설정</div>
+      <JuiceFont isBold> JUICE</JuiceFont>
+      <JuiceFont isBold> 설정 </JuiceFont>
     </div>
   );
 };
