--- src/components/line.tsx
@@ -1,7 +1,0 @@
-const Line = () => {
-  return (
-    <hr className="my-6 border-gray-200 sm:mx-auto dark:border-gray-700 lg:my-8" />
-  );
-};
-
-export default Line;
